refactor: 정렬 버튼 생성 함수 분리
